requirements.txt
pdfplumber
python-docx
pillow
pytesseract
pdf2image
openai-agents[viz]
python-dotenv